shorten rust_analyzer workspace names
cargo build script runner is not currently sufficiently long path aware,
and will fail on win32 even with a bazel temp dir of C:\bzl because it's
trying to set the current directory to a long path without
canonicalizing it properly. Rust std::fs::canonicalize won't work for
this because of reasons and I haven't yet published the crate which does
it properly, so for now, squeak under the path max limit by shortening
the names a bit.